refactor(migrateUser): add migrate_user requests to dedicated file
